review and fix bug
