+++ rolling_update.go
@@ -19,10 +19,8 @@ package pegasus
 
 import (
 	"errors"
-	"pegasus-cluster-cli/client"
-	"time"
-
 	log "github.com/sirupsen/logrus"
+	"pegasus-cluster-cli/client"
 )
 
 // RollingUpdateNodes implements the rolling-update command.
@@ -97,7 +95,7 @@ func RollingUpdateNodes(cluster string, deploy Deployment, metaList string, node
 func rollingUpdateNode(deploy Deployment, metaClient MetaClient, node *ReplicaNode) error {
 	log.Printf("rolling update replica node \"%s\" [%s]", node.Name(), node.IPPort())
 
-	// TODO(wutao): add a log
+	log.Print("Set lb.add_secondary_max_count_for_one_node to 0...")
 	if _, err := metaClient.RemoteCommand("meta.lb.add_secondary_max_count_for_one_node", "0"); err != nil {
 		return err
 	}
@@ -106,74 +104,21 @@ func rollingUpdateNode(deploy Deployment, metaClient MetaClient, node *ReplicaNo
 		return err
 	}
 
-	log.Print("Downgrading replicas on node...")
-	c := 0
-	var gpids []string
-	var err error
-	fin, err := waitFor(func() (bool, error) {
-		if c%10 == 0 {
-			gpids, err = metaClient.Downgrade(node.IPPort())
-			if err != nil {
-				return false, err
-			}
-			log.Print("Sent downgrade propose")
-		}
-		if err := node.updateInfo(metaClient); err != nil {
-			return false, err
-		}
-		log.Printf("Still %d primary replicas left on %s", node.PrimaryCount, node.IPPort())
-		c++
-		return node.PrimaryCount == 0, nil
-	}, time.Second, 28)
+	gpids, err := downgradeReplicasOnNode(metaClient, node)
 	if err != nil {
 		return err
 	}
-	if fin {
-		log.Print("Downgrade done")
-	} else {
-		log.Print("Downgrade timeout")
-	}
-	time.Sleep(time.Second)
 
-	// TODO: Check replicas closed on node here
 	remoteCmdClient := client.NewRemoteCmdClient(node.IPPort())
-	c = 0
-	log.Print("Checking replicas closed on node...")
-	fin, err = waitFor(func() (bool, error) {
-		if c%10 == 0 {
-			log.Print("Send kill_partition commands to node...")
-			for _, gpid := range gpids {
-				if _, err := remoteCmdClient.KillPartition(gpid); err != nil {
-					return false, err
-				}
-			}
-			log.Printf("Sent to %d partitions.", len(gpids))
-		}
-		counters, err := remoteCmdClient.GetPerfCounters(".*replica(Count)")
-		if err != nil {
-			return false, err
-		}
-		count := 0
-		for _, counter := range counters {
-			count += int(counter.Value)
-		}
-		log.Printf("Still %d replicas not closed on %s", count, node.IPPort())
-		c++
-		return count == 0, nil
-	}, time.Second, 28)
-	if err != nil {
+	if err := CloseReplicasOnNode(remoteCmdClient, node, gpids); err != nil {
 		return err
-	}
-	if fin {
-		log.Print("Close done.")
-	} else {
-		log.Print("Close timeout.")
 	}
 
 	if _, err := remoteCmdClient.Call("flush_log", nil); err != nil {
 		return err
 	}
 
+	log.Print("Set lb.add_secondary_max_count_for_one_node to 100...")
 	if _, err := metaClient.RemoteCommand("meta.lb.add_secondary_max_count_for_one_node", "100"); err != nil {
 		return err
 	}
@@ -184,68 +129,13 @@ func rollingUpdateNode(deploy Deployment, metaClient MetaClient, node *ReplicaNo
 	}
 	log.Print("Rolling update by deployment done")
 
-	log.Printf("Wait %s to become alive...", node.IPPort())
-	if _, err := waitFor(func() (bool, error) {
-		if err := node.updateInfo(metaClient); err != nil {
-			return false, err
-		}
-		return node.Status == "ALIVE", nil
-	}, time.Second, 0); err != nil {
+	if _, err := waitNodeAlive(node, metaClient); err != nil {
 		return err
 	}
 
-	log.Printf("Wait %s to become healthy...", node.IPPort())
-	if _, err := waitFor(func() (bool, error) {
-		infos, err := metaClient.ListTableHealthInfos()
-		if err != nil {
-			return false, err
-		}
-		count := 0
-		for _, info := range infos {
-			if info.PartitionCount != info.FullyHealthy {
-				count++
-			}
-		}
-		if count != 0 {
-			log.Printf("Cluster not healthy, unhealthy app count %d", count)
-			return false, nil
-		}
-		log.Print("Cluster becomes healthy")
-		return true, nil
-	}, time.Duration(10)*time.Second, 0); err != nil {
+	if _, err := waitNodeHealthy(node, metaClient); err != nil {
 		return err
 	}
-
-	return nil
-}
-
-func migratePrimariesOutOfNode(metaClient MetaClient, node *ReplicaNode) error {
-	c := 0
-	log.Print("migrating primary replicas out of node...")
-	fin, err := waitFor(func() (bool, error) {
-		if c%10 == 0 {
-			if err := metaClient.Migrate(node.IPPort()); err != nil {
-				return false, err
-			}
-			log.Print("proposed primaries migration to MetaServer")
-		}
-		if err := node.updateInfo(metaClient); err != nil {
-			return false, err
-		}
-		log.Printf("Still %d primary replicas left on %s", node.PrimaryCount, node.IPPort())
-		c++
-		return node.PrimaryCount == 0, nil
-	}, time.Second, 28)
-
-	if err != nil {
-		return err
-	}
-	if fin {
-		log.Print("migrate done")
-	} else {
-		log.Print("migrate timeout")
-	}
-	time.Sleep(time.Second)
 
 	return nil
 }
